+++ MemfaultPackages/bort/src/main/java/com/memfault/bort/metrics/DiskActivityStorage.kt
@@ -1,6 +1,8 @@
 package com.memfault.bort.metrics
 
 import android.content.SharedPreferences
+import com.memfault.bort.DiskActivityPrefs
+import com.memfault.bort.PREFERENCE_DISK_ACTIVITY
 import com.memfault.bort.boot.LinuxBootId
 import com.memfault.bort.shared.SerializedCachedPreferenceKeyProvider
 import com.squareup.anvil.annotations.ContributesBinding
@@ -13,11 +15,11 @@ interface DiskActivityStorage {
 
 @ContributesBinding(SingletonComponent::class, boundType = DiskActivityStorage::class)
 class RealDiskActivityStorage @Inject constructor(
-    sharedPreferences: SharedPreferences,
+    @DiskActivityPrefs sharedPreferences: SharedPreferences,
     readBootId: LinuxBootId,
 ) : DiskActivityStorage, SerializedCachedPreferenceKeyProvider<DiskActivity>(
     sharedPreferences,
     DiskActivity.EMPTY.copy(bootId = readBootId()),
     DiskActivity.serializer(),
-    "DISK_ACTIVITY",
+    PREFERENCE_DISK_ACTIVITY,
 )